redid project from scratch using stackViews
