--- PasswordTextField/PasswordTextField/PasswordField.swift
@@ -1,232 +1,0 @@
-//
-//  PasswordField.swift
-//  PasswordTextField
-//
-//  Created by Ben Gohlke on 6/26/19.
-//  Copyright © 2019 Lambda School. All rights reserved.
-//
-
-import UIKit
-
-enum PassStrengthColor: String {
-    case none = "No Password"
-    case weak = "Weak Password"
-    case medium = "Could Be Stronger"
-    case strong = "Strong Password"
-}
-
-//@IBDesignable
-class PasswordField: UIControl {
-    
-    // Public API - these properties are used to fetch the final password and strength values
-    private (set) var password: String = ""
-    private (set) var showPassword: Bool = false
-    private (set) var passwordStrength: PassStrengthColor = .weak
-    
-    
-    private let standardMargin: CGFloat = 8.0
-    private let textFieldContainerHeight: CGFloat = 50.0
-    private let textFieldMargin: CGFloat = 6.0
-    private let colorViewSize: CGSize = CGSize(width: 60.0, height: 5.0)
-    
-    private let labelTextColor = UIColor(hue: 233.0/360.0, saturation: 16/100.0, brightness: 41/100.0, alpha: 1)
-    private let labelFont = UIFont.systemFont(ofSize: 14.0, weight: .semibold)
-    
-    private let textFieldBorderColor = UIColor(hue: 208/360.0, saturation: 80/100.0, brightness: 94/100.0, alpha: 1)
-    private let bgColor = UIColor(hue: 0, saturation: 0, brightness: 97/100.0, alpha: 1)
-    
-    // States of the password strength indicators
-    private let unusedColor = UIColor(hue: 210/360.0, saturation: 5/100.0, brightness: 86/100.0, alpha: 1)
-    private let weakColor = UIColor(hue: 0/360, saturation: 60/100.0, brightness: 90/100.0, alpha: 1)
-    private let mediumColor = UIColor(hue: 39/360.0, saturation: 60/100.0, brightness: 90/100.0, alpha: 1)
-    private let strongColor = UIColor(hue: 132/360.0, saturation: 60/100.0, brightness: 75/100.0, alpha: 1)
-    
-    private var titleLabel: UILabel = UILabel()
-    private var textField: UITextField = UITextField()
-    private var showHideButton: UIButton = UIButton()
-    private var weakView: UIView = UIView()
-    private var mediumView: UIView = UIView()
-    private var strongView: UIView = UIView()
-    private var strengthDescriptionLabel: UILabel = UILabel()
-    
-    func setup() {
-      
-        // Lay out your subviews here
-        // Background
-        layer.cornerRadius = 10
-        backgroundColor = bgColor
-        
-        //Title Label & constraints
-        addSubview(titleLabel)
-        titleLabel.translatesAutoresizingMaskIntoConstraints = false
-        titleLabel.text = "ENTER PASSWORD"
-        titleLabel.font = labelFont
-        titleLabel.textColor = labelTextColor
-        titleLabel.bottomAnchor.constraint(equalTo: self.topAnchor, constant: standardMargin).isActive = true
-        titleLabel.leadingAnchor.constraint(equalTo: self.leadingAnchor, constant: standardMargin).isActive = true
-        titleLabel.trailingAnchor.constraint(equalTo: self.trailingAnchor, constant: -standardMargin).isActive = true
-        
-        
-        // Text Field & constraints
-        addSubview(textField)
-        textField.translatesAutoresizingMaskIntoConstraints = false
-        textField.topAnchor.constraint(equalTo: titleLabel.bottomAnchor, constant: textFieldMargin).isActive = true
-        textField.leadingAnchor.constraint(equalTo: leadingAnchor, constant: textFieldMargin).isActive = true
-        textField.trailingAnchor.constraint(equalTo: trailingAnchor, constant: -textFieldMargin).isActive = true
-        textField.heightAnchor.constraint(equalToConstant: textFieldContainerHeight).isActive = true
-        textField.borderStyle = .roundedRect
-        textField.layer.borderColor = textFieldBorderColor.cgColor
-        textField.layer.borderWidth = 3.0
-        textField.layer.cornerRadius = 10
-        textField.isSecureTextEntry = true
-        textField.placeholder = "Enter Your Password"
-        textField.rightView = showHideButton
-        textField.rightViewMode = .always
-        
-     
-        // show hide button
-        addSubview(showHideButton)
-        showHideButton.translatesAutoresizingMaskIntoConstraints = false
-        showHideButton.topAnchor.constraint(equalTo: textField.topAnchor, constant: 5).isActive = true
-        showHideButton.trailingAnchor.constraint(equalTo: textField.trailingAnchor, constant: -10).isActive = true
-        showHideButton.bottomAnchor.constraint(equalTo: textField.bottomAnchor, constant: -5).isActive = true
-        showHideButton.setImage(UIImage(named: "eyes-closed"), for: .normal)
-        showHideButton.addTarget(self, action: #selector(hideShowBtnTapped), for: .touchUpInside)
-        
-        // Weak Strength Labels below textfield
-        addSubview(weakView)
-        weakView.translatesAutoresizingMaskIntoConstraints = false
-        weakView.backgroundColor = unusedColor
-        weakView.topAnchor.constraint(equalTo: textField.bottomAnchor, constant: 15).isActive = true
-        weakView.leadingAnchor.constraint(equalTo: leadingAnchor, constant: textFieldMargin).isActive = true
-        weakView.heightAnchor.constraint(equalToConstant: colorViewSize.height).isActive = true
-        weakView.widthAnchor.constraint(equalToConstant: colorViewSize.width).isActive = true
-        
-         // Weak Strength Labels
-        addSubview(mediumView)
-        mediumView.translatesAutoresizingMaskIntoConstraints = false
-        mediumView.backgroundColor = mediumColor
-        mediumView.topAnchor.constraint(equalTo: textField.bottomAnchor, constant: 15).isActive = true
-        mediumView.leadingAnchor.constraint(equalTo: weakView.trailingAnchor, constant: 2).isActive = true
-        mediumView.heightAnchor.constraint(equalToConstant: colorViewSize.height).isActive = true
-        mediumView.widthAnchor.constraint(equalToConstant: colorViewSize.width).isActive = true
-        
-        // Strong Strength Labels
-        addSubview(strongView)
-        strongView.translatesAutoresizingMaskIntoConstraints = false
-        strongView.backgroundColor = strongColor
-        strongView.topAnchor.constraint(equalTo: textField.bottomAnchor, constant: 15).isActive = true
-        strongView.leadingAnchor.constraint(equalTo: mediumView.trailingAnchor, constant: 2).isActive = true
-        strongView.heightAnchor.constraint(equalToConstant: colorViewSize.height).isActive = true
-        strongView.widthAnchor.constraint(equalToConstant: colorViewSize.width).isActive = true
-        
-        
-        // Strength Description Labels
-        addSubview(strengthDescriptionLabel)
-        strengthDescriptionLabel.translatesAutoresizingMaskIntoConstraints = false
-        strengthDescriptionLabel.textColor = labelTextColor
-        strengthDescriptionLabel.font = labelFont
-        strengthDescriptionLabel.text = PassStrengthColor.weak.rawValue
-        strengthDescriptionLabel.topAnchor.constraint(equalTo: textField.bottomAnchor, constant: 8).isActive = true
-        strengthDescriptionLabel.leadingAnchor.constraint(equalTo: strongView.trailingAnchor, constant: 5).isActive = true
-        strengthDescriptionLabel.trailingAnchor.constraint(equalTo: trailingAnchor, constant: textFieldMargin).isActive = true
-
-        
-    }
-    
-    //show hide button
-    @objc func hideShowBtnTapped() {
-        showPassword.toggle()
-        if showPassword {
-            textField.isSecureTextEntry = false
-            showHideButton.setImage(UIImage(named: "eyes-open"), for: .normal)
-        } else {
-            textField.isSecureTextEntry = true
-            showHideButton.setImage(UIImage(named: "eyes-closed"), for: .normal)
-        }
-
-        }
-    
-    func passwordStrength(enteredPassword: String) {
-        switch enteredPassword.count {
-        case 0:
-            strengthDescriptionLabel.text = PassStrengthColor.none.rawValue
-            weakView.backgroundColor = unusedColor
-            mediumView.backgroundColor = unusedColor
-            strongView.backgroundColor = unusedColor
-        case 1...6:
-            strengthDescriptionLabel.text = PassStrengthColor.weak.rawValue
-            weakView.backgroundColor = unusedColor
-            mediumView.backgroundColor = unusedColor
-            strongView.backgroundColor = unusedColor
-            
-        case 7...11:
-            strengthDescriptionLabel.text = PassStrengthColor.medium.rawValue
-            weakView.backgroundColor = unusedColor
-            mediumView.backgroundColor = mediumColor
-            strongView.backgroundColor = unusedColor
-        case 12...20:
-            strengthDescriptionLabel.text = PassStrengthColor.strong.rawValue
-            weakView.backgroundColor = unusedColor
-            mediumView.backgroundColor = mediumColor
-            strongView.backgroundColor = strongColor
-        
-        default:
-            strengthDescriptionLabel.text = "Please try again"
-        }
-        
-        if enteredPassword.count == 1 {
-            weakView.performFlare()
-        } else if enteredPassword.count == 7 {
-            weakView.performFlare()
-            mediumView.performFlare()
-        } else if enteredPassword.count == 12 {
-            weakView.performFlare()
-            mediumView.performFlare()
-            strongView.performFlare()
-        }
-        
-    }
-         
-    
-    required init?(coder aDecoder: NSCoder) {
-        super.init(coder: aDecoder)
-        setup()
-        textField.delegate = self
-    }
-}
-
-extension PasswordField: UITextFieldDelegate {
-    func textField(_ textField: UITextField, shouldChangeCharactersIn range: NSRange, replacementString string: String) -> Bool {
-        let oldText = textField.text!
-        let stringRange = Range(range, in: oldText)!
-        let newText = oldText.replacingCharacters(in: stringRange, with: string)
-        // TODO: send new text to the determine strength method
-        passwordStrength(enteredPassword: newText)
-        return true
-    }
-    
-    func textFieldShouldReturn(_ textField: UITextField) -> Bool {
-        if let enteredPassword = textField.text {
-            password = enteredPassword
-            sendActions(for: .valueChanged)
-            }
-            weakView.performFlare()
-            mediumView.performFlare()
-            strongView.performFlare()
-            textField.resignFirstResponder()
-            return true
-        }
-}
-
-extension UIView {
-  // "Flare view" animation sequence
-  func performFlare() {
-    func flare()   { transform = CGAffineTransform(scaleX: 1.6, y: 1.6) }
-    func unflare() { transform = .identity }
-    
-    UIView.animate(withDuration: 0.3,
-                   animations: { flare() },
-                   completion: { _ in UIView.animate(withDuration: 0.1) { unflare() }})
-  }
-}
