+++ PasswordTextField/PasswordTextField/Resources/Assets.xcassets/AppIcon.appiconset/Contents.json
