+++ PasswordTextField/PasswordTextField/Resources/Assets.xcassets/Contents.json
