+++ PasswordTextField/PasswordTextField/Resources/Assets.xcassets/eyes-closed.imageset/Contents.json
