+++ PasswordTextField/PasswordTextField/Resources/Assets.xcassets/eyes-closed.imageset/eyes-closed.png
