+++ PasswordTextField/PasswordTextField/Resources/Assets.xcassets/eyes-open.imageset/Contents.json
